style: Update navbar links and remove unused GitHub icon
Remove the unused GitHub icon and update the navbar links in the navbar component. This improves the visual appearance and functionality of the navbar.

style: Update background color in Layout.astro

Update the background color in the Layout.astro file. This change modifies the styling of the layout to enhance the overall design.

style: Remove empty style block in Layout.astro and index.astro

Remove the empty style block in the Layout.astro and index.astro files. This cleanup improves the code readability and removes unnecessary code.

chore: Update npm dependencies and add @midudev/tailwind-animations package

Update the npm dependencies and add the @midudev/tailwind-animations package. This ensures that the project is using the latest dependencies and adds new animations capabilities.

style: Fix typo in navbar link text

Fix a typo in the navbar link text. This correction ensures that the link text is spelled correctly.